ADD: prefixes related functions
- add is_ip_v6() to utils
- prefix_encoder -> prefix encoder and decoder
- prefix_weight_calculator -> calculate prefix weight

prefix_encoder:
This script contains method of encoding prefix (network + subnet)
to integer. This data form works better with prefix weight
calculation.

prefix_weight_calculator:
Caculate prefix weight by number of addresses it contains. Keep in
mind that if there is a prefix which is a supernet for others. The
weigth of super prefix would be subtract by its smaller network.