Fix overrepresentation of peer ASes that are present at multiple collectors
While we average scores for a single peer AS that has multiple peering
sessions with a single collector, we failed to do so for ASes that peer
with multiple collectors.

This leads to an overrepresentation of the AS in the final score
calculation, since it will have multiple entries in the score list,
which will be trimmed and averaged.

We fix this by keeping track of which peer ASN provided a score when
merging the data from multiple collectors and averaging the scores in
case an ASN provides more than one.