Fix hegemony scores higher than 1
There are some instances of hegemony scores higher than 1 caused by
technically invalid AS paths.

A score higher than 1 can occur if an on-path AS shows up more often
than the origin AS, which is used as the base for normalization.
Normally this should never happen since we remove AS path prepending,
but there are some non-compliant paths of the form A B A, which we did
not handle before.

This PR changes the deduplication to take into account all ASes seen in
the path, not just the previous one.

Fixes #4.