Fix waiting for dead collectors
If there is no data available for a collecor, we want to ignore it for
the rest of the analysis round.

However, `loaded_bcscore` can never be None, since it is always
initialized in `BCSSCORELoader.prepare_load_data()` and returned by
`DataLoader.load_data()`. For this to work correctly, we need to check
if it is empty instead.